Extracted interface for Neo4jPack
To make it easier to plug different version that, for example, supports
new types, like point, geometry or date.
