+++ community/bolt/src/test/java/org/neo4j/bolt/v1/messaging/BoltResponseMessageTest.java
@@ -226,10 +226,11 @@ public class BoltResponseMessageTest
     private <T extends ResponseMessage> T serializeAndDeserialize( T msg ) throws IOException
     {
         RecordingByteChannel channel = new RecordingByteChannel();
+        Neo4jPack neo4jPack = new Neo4jPackV1();
         BoltResponseMessageReader reader = new BoltResponseMessageReader(
-                new Neo4jPack.Unpacker( new BufferedChannelInput( 16 ).reset( channel ) ) );
+                neo4jPack.newUnpacker( new BufferedChannelInput( 16 ).reset( channel ) ) );
         BoltResponseMessageWriter writer = new BoltResponseMessageWriter(
-                new Neo4jPack.Packer( new BufferedChannelOutput( channel ) ), NO_BOUNDARY_HOOK,
+                neo4jPack.newPacker( new BufferedChannelOutput( channel ) ), NO_BOUNDARY_HOOK,
                 NullBoltMessageLogger.getInstance() );
 
         msg.dispatch( writer );
