+++ community/bolt/src/test/java/org/neo4j/bolt/v1/messaging/util/MessageMatchers.java
@@ -40,6 +40,7 @@ import org.neo4j.bolt.v1.messaging.BoltResponseMessageReader;
 import org.neo4j.bolt.v1.messaging.BoltResponseMessageRecorder;
 import org.neo4j.bolt.v1.messaging.BoltResponseMessageWriter;
 import org.neo4j.bolt.v1.messaging.Neo4jPack;
+import org.neo4j.bolt.v1.messaging.Neo4jPackV1;
 import org.neo4j.bolt.v1.messaging.RecordingByteChannel;
 import org.neo4j.bolt.v1.messaging.message.FailureMessage;
 import org.neo4j.bolt.v1.messaging.message.IgnoredMessage;
@@ -50,13 +51,13 @@ import org.neo4j.bolt.v1.messaging.message.SuccessMessage;
 import org.neo4j.bolt.v1.packstream.BufferedChannelInput;
 import org.neo4j.bolt.v1.packstream.BufferedChannelOutput;
 import org.neo4j.bolt.v1.transport.integration.TestNotification;
-import org.neo4j.kernel.impl.util.BaseToObjectValueWriter;
 import org.neo4j.cypher.result.QueryResult;
 import org.neo4j.graphdb.Node;
 import org.neo4j.graphdb.Notification;
 import org.neo4j.graphdb.Relationship;
 import org.neo4j.graphdb.spatial.Point;
 import org.neo4j.kernel.api.exceptions.Status;
+import org.neo4j.kernel.impl.util.BaseToObjectValueWriter;
 import org.neo4j.kernel.impl.util.HexPrinter;
 import org.neo4j.values.AnyValue;
 import org.neo4j.values.storable.CoordinateReferenceSystem;
@@ -295,30 +296,33 @@ public class MessageMatchers
 
     public static byte[] serialize( RequestMessage... messages ) throws IOException
     {
-        final RecordingByteChannel rawData = new RecordingByteChannel();
-        final BoltRequestMessageWriter packer = new BoltRequestMessageWriter( new Neo4jPack.Packer( new
-                BufferedChannelOutput( rawData ) ), NO_BOUNDARY_HOOK );
+        RecordingByteChannel rawData = new RecordingByteChannel();
+        Neo4jPack neo4jPack = new Neo4jPackV1();
+        Neo4jPack.Packer packer = neo4jPack.newPacker( new BufferedChannelOutput( rawData ) );
+        BoltRequestMessageWriter writer = new BoltRequestMessageWriter( packer, NO_BOUNDARY_HOOK );
 
         for ( RequestMessage message : messages )
         {
-            packer.write( message );
+            writer.write( message );
         }
-        packer.flush();
+        writer.flush();
 
         return rawData.getBytes();
     }
 
     public static byte[] serialize( ResponseMessage... messages ) throws IOException
     {
-        final RecordingByteChannel rawData = new RecordingByteChannel();
-        final BoltResponseMessageWriter packer = new BoltResponseMessageWriter( new Neo4jPack.Packer( new
-                BufferedChannelOutput( rawData ) ), NO_BOUNDARY_HOOK, NullBoltMessageLogger.getInstance() );
+        RecordingByteChannel rawData = new RecordingByteChannel();
+        Neo4jPack neo4jPack = new Neo4jPackV1();
+        Neo4jPack.Packer packer = neo4jPack.newPacker( new BufferedChannelOutput( rawData ) );
+        BoltResponseMessageWriter writer = new BoltResponseMessageWriter( packer, NO_BOUNDARY_HOOK,
+                NullBoltMessageLogger.getInstance() );
 
         for ( ResponseMessage message : messages )
         {
-            message.dispatch( packer );
+            message.dispatch( writer );
         }
-        packer.flush();
+        writer.flush();
 
         return rawData.getBytes();
     }
@@ -370,14 +374,18 @@ public class MessageMatchers
 
     private static BoltRequestMessageReader requestReader( byte[] bytes )
     {
-        return new BoltRequestMessageReader(
-                new Neo4jPack.Unpacker( new BufferedChannelInput( 128 ).reset( new ArrayByteChannel( bytes ) ) ) );
+        BufferedChannelInput input = new BufferedChannelInput( 128 );
+        input.reset( new ArrayByteChannel( bytes ) );
+        Neo4jPack neo4jPack = new Neo4jPackV1();
+        return new BoltRequestMessageReader( neo4jPack.newUnpacker( input ) );
     }
 
     private static BoltResponseMessageReader responseReader( byte[] bytes )
     {
-        return new BoltResponseMessageReader(
-                new Neo4jPack.Unpacker( new BufferedChannelInput( 128 ).reset( new ArrayByteChannel( bytes ) ) ) );
+        BufferedChannelInput input = new BufferedChannelInput( 128 );
+        input.reset( new ArrayByteChannel( bytes ) );
+        Neo4jPack neo4jPack = new Neo4jPackV1();
+        return new BoltResponseMessageReader( neo4jPack.newUnpacker( input ) );
     }
 
 }
