+++ community/bolt/src/test/java/org/neo4j/bolt/v1/transport/BoltMessagingProtocolV1HandlerTest.java
@@ -31,6 +31,7 @@ import java.util.Objects;
 
 import org.neo4j.bolt.BoltChannel;
 import org.neo4j.bolt.transport.TransportThrottleGroup;
+import org.neo4j.bolt.v1.messaging.Neo4jPackV1;
 import org.neo4j.bolt.v1.runtime.BoltStateMachine;
 import org.neo4j.bolt.v1.runtime.BoltWorker;
 import org.neo4j.bolt.v1.runtime.SynchronousBoltWorker;
@@ -61,8 +62,9 @@ public class BoltMessagingProtocolV1HandlerTest
         when( boltChannel.rawChannel() ).thenReturn( outputChannel );
 
         BoltStateMachine machine = mock( BoltStateMachine.class );
-        BoltMessagingProtocolV1Handler protocol = new BoltMessagingProtocolV1Handler(
-                boltChannel, new SynchronousBoltWorker( machine ), TransportThrottleGroup.NO_THROTTLE, NullLogService.getInstance() );
+        BoltMessagingProtocolV1Handler protocol = new BoltMessagingProtocolV1Handler( boltChannel, new Neo4jPackV1(),
+                new SynchronousBoltWorker( machine ), TransportThrottleGroup.NO_THROTTLE,
+                NullLogService.getInstance() );
         verify( outputChannel ).alloc();
 
         // And given inbound data that'll explode when the protocol tries to interpret it
@@ -94,8 +96,9 @@ public class BoltMessagingProtocolV1HandlerTest
         BoltChannel boltChannel = mock( BoltChannel.class );
         when( boltChannel.rawChannel() ).thenReturn( outputChannel );
 
-        BoltMessagingProtocolV1Handler protocol = new BoltMessagingProtocolV1Handler(
-                boltChannel, new SynchronousBoltWorker( machine ), TransportThrottleGroup.NO_THROTTLE, NullLogService.getInstance() );
+        BoltMessagingProtocolV1Handler protocol = new BoltMessagingProtocolV1Handler( boltChannel, new Neo4jPackV1(),
+                new SynchronousBoltWorker( machine ), TransportThrottleGroup.NO_THROTTLE,
+                NullLogService.getInstance() );
         protocol.close();
 
         verify( machine ).close();
@@ -115,8 +118,8 @@ public class BoltMessagingProtocolV1HandlerTest
         BoltChannel boltChannel = mock( BoltChannel.class );
         when( boltChannel.rawChannel() ).thenReturn( outputChannel );
 
-        BoltMessagingProtocolV1Handler protocol = new BoltMessagingProtocolV1Handler(
-                boltChannel, mock( BoltWorker.class ), TransportThrottleGroup.NO_THROTTLE, logService );
+        BoltMessagingProtocolV1Handler protocol = new BoltMessagingProtocolV1Handler( boltChannel, new Neo4jPackV1(),
+                mock( BoltWorker.class ), TransportThrottleGroup.NO_THROTTLE, logService );
 
         protocol.handle( mock( ChannelHandlerContext.class ), data );
