+++ community/bolt/src/test/java/org/neo4j/bolt/v1/transport/socket/SocketTransportHandlerTest.java
@@ -31,6 +31,7 @@ import org.neo4j.bolt.transport.BoltMessagingProtocolHandler;
 import org.neo4j.bolt.transport.BoltProtocolHandlerFactory;
 import org.neo4j.bolt.transport.SocketTransportHandler;
 import org.neo4j.bolt.transport.TransportThrottleGroup;
+import org.neo4j.bolt.v1.messaging.Neo4jPackV1;
 import org.neo4j.bolt.v1.runtime.BoltStateMachine;
 import org.neo4j.bolt.v1.runtime.SynchronousBoltWorker;
 import org.neo4j.bolt.v1.transport.BoltMessagingProtocolV1Handler;
@@ -203,7 +204,7 @@ public class SocketTransportHandlerTest
         BoltProtocolHandlerFactory handlerFactory = ( version, channel ) ->
         {
             assertEquals( 1, version );
-            return new BoltMessagingProtocolV1Handler( channel, new SynchronousBoltWorker( machine ),
+            return new BoltMessagingProtocolV1Handler( channel, new Neo4jPackV1(), new SynchronousBoltWorker( machine ),
                     TransportThrottleGroup.NO_THROTTLE, NullLogService.getInstance() );
         };
